# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "\"赚圈圈API 返回的goods如果其中的goodsin为yes代表它可以进行交易，即：未下架\"",
        "title": "miniproject",
        "termsOfService": "http://swagger.io/terrms/",
        "contact": {
            "name": "yyj",
            "email": "2105753640@qq.com"
        },
        "version": "1.0.0"
    },
    "host": "119.3.133.235:8080",
    "basePath": "/api/vi",
    "paths": {
...
        "/money/goods/addition": {
            "post": {
                "description": "\"新增一个商品时的api\"",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Good"
                ],
                "summary": "\"上架商品\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "标题",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分区",
                        "name": "zone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "价格",
                        "name": "price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "详情",
                        "name": "summary",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "商品图二进制文件",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "联系方式二进制文件",
                        "name": "way",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"upload successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/goods/comment": {
            "post": {
                "description": "\"用户做出评价，点击评价时的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "\"用户对某个商品的评论\"",
                "parameters": [
                    {
                        "description": "评论",
                        "name": "comment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Description"
                        }
                    },
                    {
                        "type": "string",
                        "description": "商品编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"give successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/goods/comments": {
            "get": {
                "description": "\"商品详情页点击评价时的api \"scores\":所有分值情况, \"infor\":\"评论信息以及学号\"\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comment"
                ],
                "summary": "\"获取某个商品的所有评论\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "infor\":[]tables.Comment,\"score\":All}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"err\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/goods/deletion": {
            "delete": {
                "description": "\"下架商品的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Good"
                ],
                "summary": "\"商家下架商品\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"delete successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/money/goods/feedback": {
            "post": {
                "description": "\"举报的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Good"
                ],
                "summary": "\"接收举报\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "只需上传用户勾选的个数 内容不需要",
                        "name": "reasonNum",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"举报成功!\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/money/my/cancellation": {
            "post": {
                "description": "\"在购物车里取消收藏的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Star"
                ],
                "summary": "\"取消收藏\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"cancel successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/money/new_star": {
            "patch": {
                "description": "\"添加至购物车时的api\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Star"
                ],
                "summary": "\"新增收藏\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "商品的编号",
                        "name": "goodsid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "msg\":\"add successfully\" \"msg\":\"你已经收藏过该商品了\"}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "msg\":\"error happened in server\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Description": {
            "type": "object",
            "required": [
                "comment",
                "score"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
...
    }
}